Move manager logic to process trait
