+++ ProcessMaker/Traits/ProcessTrait.php
@@ -3,6 +3,7 @@
 namespace ProcessMaker\Traits;
 
 use ProcessMaker\Models\ProcessVersion;
+use ProcessMaker\Models\User;
 use ProcessMaker\Nayra\Contracts\Storage\BpmnDocumentInterface;
 use ProcessMaker\Repositories\BpmnDocument;
 
@@ -45,5 +46,61 @@ trait ProcessTrait
         $document = new BpmnDocument($this);
         $document->loadXML($this->bpmn);
         return $document;
+    }
+    
+    /**
+     * Set a value on the properties json column
+     *
+     * @param string $name
+     * @param mixed $value
+     * @return void
+     */
+    private function setProperty($name, $value)
+    {
+        $properties = $this->properties;
+        $properties[$name] = $value;
+        $this->properties = $properties;
+    }
+
+    /**
+     * Get a value from the properties json column
+     *
+     * @param string $name
+     * @return mixed
+     */
+    private function getProperty($name)
+    {
+       return isset($this->properties[$name]) ? $this->properties[$name] : null;
+    }
+
+    /**
+     * Set the manager id
+     *
+     * @param int $value
+     * @return void
+     */
+    public function setManagerIdAttribute($value)
+    {
+        $this->setProperty('manager_id', $value);
+    }
+    
+    /**
+     * Get the the manager id
+     *
+     * @return int|null
+     */
+    public function getManagerIdAttribute()
+    {
+        return $this->getProperty('manager_id');
+    }
+    
+    /**
+     * Get the process manager
+     *
+     * @return User|null
+     */
+    public function manager()
+    {
+        return $this->belongsTo(User::class, 'manager_id');
     }
 }
